+++ packages/e2e.generated.catchall/src/generatedNoCheck/index.ts
@@ -22,7 +22,11 @@ export {
   CollateralConcernList,
   EsongInterfaceA,
   FooInterface,
+  TestAsTypeIltInterface,
   InterfaceNoProps,
+  NihalbCastingInterfaceB,
+  NihalbCastingInterfaceTypeA,
+  NihalbCastingLinkedInterfaceTypeA,
   OsdkTestInterface,
 } from './ontology/interfaces.js';
 export * as $Interfaces from './ontology/interfaces.js';
@@ -42,6 +46,10 @@ export {
   MnayanOsdkMediaObject,
   MtaBus,
   NbaPlayer,
+  NihalbCastingLinkedObjectTypeA,
+  NihalbCastingObjectTypeA,
+  NihalbCastingObjectTypeB,
+  NihalbCastingObjectTypeC,
   ObjectTypeWithAllPropertyTypes,
   OsdkTestObject,
   Person,