+++ packages/e2e.generated.catchall/src/generatedNoCheck/ontology/objects/OsdkTestObject.ts
@@ -123,7 +123,7 @@ export const OsdkTestObject = {
   type: 'object',
   apiName: 'OsdkTestObject',
   osdkMetadata: $osdkMetadata,
-  experimentalDoNotUseMetadata: {
+  internalDoNotUseMetadata: {
     rid: 'ri.ontology.main.object-type.ba4a949c-547a-45de-9c78-b772bb55acfb',
   },
-} satisfies OsdkTestObject & { experimentalDoNotUseMetadata: { rid: string } } as OsdkTestObject;
+} satisfies OsdkTestObject & { internalDoNotUseMetadata: { rid: string } } as OsdkTestObject;