+++ packages/client.unstable/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @osdk/client.unstable
 
+## 2.5.0-beta.14
+
+### Minor Changes
+
+- 9dd5a63: Intermediary links in oac
+- e62408b: Fix Maker Rid Generation
+
+## 2.5.0-beta.13
+
 ## 2.5.0-beta.12
 
 ## 2.5.0-beta.11